requirements/local.txt
-r production.txt
coverage
-e git+https://github.com/Axik/django-coverage-py3.git#egg=django-coverage-py3
django-debug-toolbar
bpython
Sphinx
dj-database-url
